+++ packages/ipydrawio-jupyter-templates/src/index.ts
@@ -14,7 +14,4 @@
   limitations under the License.
 */
 
-import '@deathbeds/ipydrawio';
-import '@deathbeds/ipydrawio-pdf';
-import '@deathbeds/ipydrawio-notebook';
-import '@deathbeds/ipydrawio-jupyter-templates';
+export * from './tokens';